+++ client/src/main.rs
@@ -279,7 +279,7 @@ async fn run(
 		db.put(IsFinalitySyncedKey, true);
 	}
 
-	let delay_sec = updater::delay_sec();
+	let delay_sec = updater::delay_sec(cfg.max_restart_delay);
 	let updater_run = updater::run(
 		version,
 		delay_sec,
@@ -294,6 +294,11 @@ async fn run(
 		}
 	}));
 
+	// Delay for maintenance restart, updater restart delay + maintenance restart delay
+	let maintenance_restart_delay = cfg
+		.maintenance_restart
+		.then_some(delay_sec + cfg.maintenance_restart_delay);
+
 	let static_config_params: MaintenanceConfig = (&cfg).into();
 	let (maintenance_sender, maintenance_receiver) = mpsc::unbounded_channel::<MaintenanceEvent>();
 	spawn_in_span(shutdown.with_cancel(maintenance::run(
@@ -302,6 +307,8 @@ async fn run(
 		static_config_params,
 		shutdown.clone(),
 		maintenance_sender,
+		restart.clone(),
+		maintenance_restart_delay,
 	)));
 
 	let channels = avail_light_core::types::ClientChannels {