+++ monitor-client/src/main.rs
@@ -1,3 +1,5 @@
+use crate::server::{get_blacklisted_peers, get_peer_count, get_peers};
+use actix_web::{web, App, HttpServer};
 use avail_light_core::{
 	data,
 	network::p2p::{self, is_multiaddr_global, OutputEvent},
@@ -11,8 +13,10 @@ use color_eyre::{eyre::Context, Result};
 use config::CliOpts;
 use libp2p::{Multiaddr, PeerId};
 use peer_monitor::PeerMonitor;
+use server::AppState;
 use std::{
 	collections::{HashMap, HashSet, VecDeque},
+	net::{Ipv4Addr, SocketAddr},
 	sync::Arc,
 	time::{Duration, SystemTime},
 };
@@ -28,12 +32,14 @@ mod bootstrap_monitor;
 mod config;
 mod peer_discovery;
 mod peer_monitor;
+mod server;
 mod types;
 
 // TODO: Add pruning logic that periodically goes through the list of servers and drops servers that were not seen for a while
 
 #[tokio::main]
 async fn main() -> Result<()> {
+	// Contains blacklisted peers
 	let server_list: Arc<Mutex<HashMap<PeerId, ServerInfo>>> =
 		Arc::new(Mutex::new(HashMap::default()));
 	let server_black_list: Arc<Mutex<HashMap<PeerId, ServerInfo>>> =
@@ -146,7 +152,7 @@ async fn main() -> Result<()> {
 	let peer_monitor = PeerMonitor::new(
 		p2p_client.clone(),
 		peer_monitor_interval,
-		server_list,
+		server_list.clone(),
 		server_black_list.clone(),
 		config_clone,
 	);
@@ -155,8 +161,35 @@ async fn main() -> Result<()> {
 		if let Err(e) = peer_monitor.start_monitoring().await {
 			error!("Peer monitor error: {e}");
 		};
-	}))
-	.await;
+	}));
+
+	let app_state = web::Data::new(AppState {
+		blacklist: server_black_list.clone(),
+		server_list: server_list.clone(),
+		pagination: config.pagination,
+	});
+
+	let server = HttpServer::new(move || {
+		App::new()
+			.app_data(app_state.clone())
+			.service(get_blacklisted_peers)
+			.service(get_peers)
+			.service(get_peer_count)
+	})
+	.bind(SocketAddr::from((
+		Ipv4Addr::UNSPECIFIED,
+		config.http_port.unwrap(),
+	)))?;
+
+	match server.run().await {
+		Ok(_) => {
+			info!("HTTP server stopped");
+		},
+		Err(e) => {
+			error!("HTTP server error: {}", e);
+		},
+	}
+
 	Ok(())
 }